+++ components-core/src/main/java/org/dllearner/algorithms/qtl/filters/KeywordBasedStatementFilter2.java
@@ -248,7 +248,7 @@ public class KeywordBasedStatementFilter2 extends Filter<Statement> {
 	public Set<Statement> getStatementsBelowThreshold(double threshold){
 		Set<Statement> statements = new HashSet<>();
 		for(Entry<Statement, Double> entry : statement2Similarity.entrySet()){
-			if(entry.getValue().doubleValue() < threshold){
+			if(entry.getValue() < threshold){
 				statements.add(entry.getKey());
 			}
 		}
